BREAKING fix: fully remove providers
This removes the provider abstraction and only supports Kubernetes from now on.

In an early design period, it seemed to make sense to support other target
systems apart from Kubernetes as well. But during development, the additional
layer of abstraction hindered the development of the currently most important
feature: Great Kubernetes support.

By removing providers, tanka can concentrate on providing the best possible
Kubernetes experience.
