+++ cmd/tk/workflow.go
@@ -10,45 +10,10 @@ import (
 	"golang.org/x/crypto/ssh/terminal"
 )
 
-func providerCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "provider",
-		Short: "interact with providers",
-	}
-	cmd.AddCommand(providerListCmd())
-
-	proCmd := prov.Cmd()
-	proCmd.Use = provName
-	cmd.AddCommand(proCmd)
-	return cmd
-}
-
-func providerListCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "list",
-		Short: "print all available providers",
-		Run: func(cmd *cobra.Command, args []string) {
-			log.Println("Available providers:", listProviders())
-		},
-	}
-	return cmd
-}
-
-func listProviders() []string {
-	keys := make([]string, len(providers))
-
-	i := 0
-	for k := range providers {
-		keys[i] = k
-		i++
-	}
-	return keys
-}
-
 func applyCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "apply",
-		Short: "[Requires Provider] apply the configuration to the target",
+		Short: "apply the configuration to the cluster",
 	}
 	cmd.Run = func(cmd *cobra.Command, args []string) {
 		raw, err := evalDict()
@@ -56,12 +21,12 @@ func applyCmd() *cobra.Command {
 			log.Fatalln("evaluating jsonnet:", err)
 		}
 
-		desired, err := prov.Reconcile(raw)
+		desired, err := kube.Reconcile(raw)
 		if err != nil {
 			log.Fatalln("reconciling:", err)
 		}
 
-		if err := prov.Apply(desired); err != nil {
+		if err := kube.Apply(desired); err != nil {
 			log.Fatalln("applying:", err)
 		}
 	}
@@ -71,7 +36,7 @@ func applyCmd() *cobra.Command {
 func diffCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "diff",
-		Short: "[Requires Provider] print differences between the configuration and the target",
+		Short: "differences between the configuration and the cluster",
 	}
 	cmd.Run = func(cmd *cobra.Command, args []string) {
 		raw, err := evalDict()
@@ -79,12 +44,12 @@ func diffCmd() *cobra.Command {
 			log.Fatalln("evaluating jsonnet:", err)
 		}
 
-		desired, err := prov.Reconcile(raw)
+		desired, err := kube.Reconcile(raw)
 		if err != nil {
 			log.Fatalln("reconciling:", err)
 		}
 
-		changes, err := prov.Diff(desired)
+		changes, err := kube.Diff(desired)
 		if err != nil {
 			log.Fatalln("diffing:", err)
 		}
@@ -103,7 +68,7 @@ func diffCmd() *cobra.Command {
 func showCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "show",
-		Short: "[Requires Provider] print the jsonnet in the target state format",
+		Short: "jsonnet as yaml",
 	}
 	cmd.Run = func(cmd *cobra.Command, args []string) {
 		raw, err := evalDict()
@@ -111,12 +76,12 @@ func showCmd() *cobra.Command {
 			log.Fatalln("evaluating jsonnet:", err)
 		}
 
-		state, err := prov.Reconcile(raw)
+		state, err := kube.Reconcile(raw)
 		if err != nil {
-			log.Fatalln("invoking provider:", err)
+			log.Fatalln("reconciling:", err)
 		}
 
-		pretty, err := prov.Fmt(state)
+		pretty, err := kube.Fmt(state)
 		if err != nil {
 			log.Fatalln("pretty printing state:", err)
 		}
